requirements.txt
python-amazon-simple-product-api
praw